refactor(cloudflare): extract shared mock utilities
- Create cloudflare-worker/src/data/common.ts with shared data:
  - titles, avatars, covers, desc, user arrays
  - defaultUser, members, fakeList
- Refactor routes/list.ts to use shared fakeList and defaultUser
- Refactor routes/settings.ts to use shared defaultUser
- Refactor data/user.ts to use shared defaultUser
- Add path alias @/ in tsconfig.json

Removed ~188 lines of duplicate code.
