+++ cloudflare-worker/tsconfig.json
@@ -8,7 +8,11 @@
     "lib": ["ES2021"],
     "types": ["@cloudflare/workers-types"],
     "resolveJsonModule": true,
-    "esModuleInterop": true
+    "esModuleInterop": true,
+    "baseUrl": ".",
+    "paths": {
+      "@/*": ["src/*"]
+    }
   },
   "include": ["src/**/*"]
 }
